requirements.txt
gunicorn
gevent
tensorflow-gpu
Flask-APScheduler
Flask-SQLAlchemy
bs4
Flask
numpy
pandas
pandas-datareader
selenium
tqdm
sklearn
yfinance
git+https://github.com/hanzhichao2000/pysentiment